Added discord commands
Using discord commands for controlling youtube connection. Easier to connect to a livechat if we already know the video id. Otherwise we need to search for the video before we can find the chat id. Also due to rate limits on the youtube side, it's better if we're prompted to connect instead of trying to proactively check